defines a separate go file for the BenchTest type and methods
