+++ test/e2e/benchmark/throughput.go
@@ -21,76 +21,6 @@ func main() {
 	}
 }
 
-type BenchTest struct {
-	*testnet.Testnet
-	manifest *testnet.Manifest
-}
-
-func NewBenchTest(name string, manifest *testnet.Manifest) (*BenchTest, error) {
-	// create a new testnet
-	testNet, err := testnet.New(name, seed,
-		testnet.GetGrafanaInfoFromEnvVar(), manifest.ChainID,
-		manifest.GetGenesisModifiers()...)
-	if err != nil {
-		return nil, err
-	}
-
-	testNet.SetConsensusParams(manifest.GetConsensusParams())
-	return &BenchTest{Testnet: testNet, manifest: manifest}, nil
-}
-
-func (b *BenchTest) SetupNodes() error {
-	testnet.NoError("failed to create genesis nodes",
-		b.CreateGenesisNodes(b.manifest.Validators,
-			b.manifest.CelestiaAppVersion, b.manifest.SelfDelegation,
-			b.manifest.UpgradeHeight, b.manifest.ValidatorResource))
-
-	// obtain the GRPC endpoints of the validators
-	gRPCEndpoints, err := b.RemoteGRPCEndpoints()
-	testnet.NoError("failed to get validators GRPC endpoints", err)
-	log.Println("validators GRPC endpoints", gRPCEndpoints)
-
-	// create txsim nodes and point them to the validators
-	log.Println("Creating txsim nodes")
-
-	err = b.CreateTxClients(b.manifest.TxClientVersion,
-		b.manifest.BlobSequences,
-		b.manifest.BlobSizes,
-		b.manifest.TxClientsResource, gRPCEndpoints)
-	testnet.NoError("failed to create tx clients", err)
-
-	// start the testnet
-	log.Println("Setting up testnet")
-	testnet.NoError("failed to setup testnet", b.Setup(
-		testnet.WithPerPeerBandwidth(b.manifest.PerPeerBandwidth),
-		testnet.WithTimeoutPropose(b.manifest.TimeoutPropose),
-		testnet.WithTimeoutCommit(b.manifest.TimeoutCommit),
-		testnet.WithPrometheus(b.manifest.Prometheus),
-	))
-	return nil
-}
-
-func (b *BenchTest) Run() error {
-	log.Println("Starting testnet")
-	err := b.Start()
-	if err != nil {
-		return fmt.Errorf("failed to start testnet: %v", err)
-	}
-
-	// once the testnet is up, start the txsim
-	log.Println("Starting tx clients")
-	err = b.StartTxClients()
-	if err != nil {
-		return fmt.Errorf("failed to start tx clients: %v", err)
-	}
-
-	// wait some time for the txsim to submit transactions
-	time.Sleep(b.manifest.TestDuration)
-
-	// TODO perhaps we can stop the nodes at this point to save resources
-	return nil
-}
-
 func E2EThroughput() error {
 	latestVersion, err := testnet.GetLatestVersion()
 	testnet.NoError("failed to get latest version", err)
